♻️ Extract Catalog endpoint host into a reusable static class
Moves Catalog's registration (DbContext, endpoint configuration,
handler registration, AddNServiceBusEndpoint) into CatalogEndpointHost
so the same code can be called from Catalog's own Program.cs and from
OmNomNom.AllInOne. Also disables assembly scanning on the endpoint
configuration: harmless in standalone (only Catalog's own handler
types are loaded) and required in AllInOne (where every endpoint's
[Handler]-marked types would otherwise be picked up by every config).
